+++ tests/src/tests/json/mod.rs
@@ -3,7 +3,7 @@ pub mod optional;
 #[cfg(test)]
 mod test {
     use axum_test::TestServer;
-    use jder_axum::response::json::{JsonResponse, JsonResponseErrorCode};
+    use jder_axum::response::json::{JsonResponse, ResponseError};
 
     use crate::router::create_server;
     use crate::router::json::RouteJsonResponseData;
@@ -74,9 +74,6 @@ mod test {
             server.post("/json").await.json::<RouteResponse>();
 
         assert_eq!(res.success, false);
-        assert_eq!(
-            res.errors[0].code,
-            JsonResponseErrorCode::Parse.to_string()
-        );
+        assert_eq!(res.errors[0].code, ResponseError::Parse.as_code());
     }
 }