+++ tests/src/tests/path.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod test {
     use axum_test::TestServer;
-    use jder_axum::response::json::{JsonResponse, JsonResponseErrorCode};
+    use jder_axum::response::json::{JsonResponse, ResponseError};
 
     use crate::router::create_server;
     use crate::router::path::RoutePathResponseData;
@@ -31,9 +31,6 @@ mod test {
             server.post("/path/123A/Name").await.json::<JsonResponse>();
 
         assert_eq!(res.success, false);
-        assert_eq!(
-            res.errors[0].code,
-            JsonResponseErrorCode::Parse.to_string()
-        );
+        assert_eq!(res.errors[0].code, ResponseError::Parse.as_code());
     }
 }