+++ tests/src/tests/request_time_limit.rs
@@ -2,7 +2,7 @@
 mod test {
     use axum_test::TestServer;
     use jder_axum::response::json::{
-        JsonResponse, JsonResponseError, JsonResponseErrorCode,
+        JsonResponse, JsonResponseError, ResponseError,
     };
 
     use crate::router::create_server;
@@ -34,6 +34,6 @@ mod test {
 
         let err: JsonResponseError = res.errors[0].clone();
 
-        assert_eq!(err.code, JsonResponseErrorCode::Timeout.as_str());
+        assert_eq!(err.code, ResponseError::Timeout.as_code());
     }
 }