+++ tests/src/tests/typed_header/mod.rs
@@ -4,7 +4,7 @@ pub mod optional;
 mod test {
     use axum::http::header;
     use axum_test::TestServer;
-    use jder_axum::response::json::{JsonResponse, JsonResponseErrorCode};
+    use jder_axum::response::json::{JsonResponse, ResponseError};
 
     use crate::router::create_server;
 
@@ -35,9 +35,6 @@ mod test {
             server.post("/typed_header").await.json::<RouteResponse>();
 
         assert_eq!(res.success, false);
-        assert_eq!(
-            res.errors[0].code,
-            JsonResponseErrorCode::Parse.to_string()
-        );
+        assert_eq!(res.errors[0].code, ResponseError::Parse.as_code());
     }
 }